Rotation
You are given two strings  and  of the same length . Your task is to convert the string  into  by doing some operations. In an operation, you can delete the first character of the string  and append any character at the end of the string. You are required to determine the minimum number of operations to convert  into .

Input format

First line: Single integer  denoting the length of the strings
Second line: String 
Third line: String 


sample input::
7
aaxaabc
aabcaax

output::
3